requirements.txt
streamlit
google-cloud-speech
google-cloud-texttospeech
moviepy
requests
numpy
google-auth
pydub